Move TextArea to text-input folder
