+++ src/components/form/ItemForm.jsx
@@ -3,7 +3,7 @@ import ImageEditList from "../gallery/ImageEditList";
 import { useState } from "react";
 import FlexImage from "../image/FlexImage";
 import ImageCarousel from "../gallery/ImageCarousel";
-import TextArea from "./TextArea";
+import TextArea from "../text-input/TextArea";
 
 const sampleImage =
   "https://i1.wp.com/jejuhydrofarms.com/wp-content/uploads/2020/05/blank-profile-picture-973460_1280.png?ssl=1";
